Retuve Live Edits
Retuve live is more functionally ready.
Added some changes to metric limits
Bug Fixes

+++ retuve/app/utils.py
@@ -1,11 +1,37 @@
 import asyncio
+import hashlib
 import json
 import os
+import secrets
 from datetime import datetime
 
+from fastapi import (
+    APIRouter,
+    Depends,
+    Header,
+    HTTPException,
+    Request,
+    Response,
+)
 from httpx import AsyncClient
 
 from retuve.funcs import retuve_run
+
+
+def generate_token():
+    return secrets.token_urlsafe(32)
+
+
+def consistent_hash(value, mod):
+    # Convert the input to a string and encode it
+    encoded_value = str(value).encode()
+    # Use a reproducible hash function (e.g., SHA-256)
+    hash_value = int(hashlib.sha256(encoded_value).hexdigest(), 16)
+    return hash_value % mod
+
+
+TOKEN_STORE = {}
+API_TOKEN_STORE = {}
 
 
 async def save_dicom_and_get_results(
@@ -13,8 +39,12 @@ async def save_dicom_and_get_results(
 ):
     # Save the DICOM file in the appropriate directory
     dicom_path = f"{live_batchdir}/{instance_id}.dcm"
-    with open(dicom_path, "wb") as f:
-        f.write(dicom_content)
+    if dicom_content is not None:
+        # if path already exists, return
+        if os.path.exists(dicom_path):
+            return
+        with open(dicom_path, "wb") as f:
+            f.write(dicom_content)
 
     # Process the DICOM
     result = await asyncio.to_thread(
@@ -107,6 +137,15 @@ async def get_sorted_dicom_images(
                             auth=auth,
                         )
                         metadata = metadata_response.json()
+
+                        # Remove files that are not multiframe US's
+                        if not (
+                            metadata.get("SOPClassUID")
+                            == "1.2.840.10008.5.1.4.1.1.3.1"
+                            and int(metadata.get("NumberOfFrames", 0)) > 1
+                        ):
+                            continue
+
                         acq_date, acq_time = metadata.get(
                             "AcquisitionDate"
                         ), metadata.get("AcquisitionTime")
@@ -132,6 +171,10 @@ async def get_sorted_dicom_images(
                     auth=auth,
                 )
 
+                # hash the main instance_id to be smaller
+                small_id = hash(instance_id) % (10**8)
+                instance_id = f"{acq_datetime.strftime('%Y-%m-%d-%H:%M:%S')} ID-{small_id}"
+
                 latest_acq_time = max(latest_acq_time, acq_datetime)
 
                 final_images_with_dates.append(
@@ -147,10 +190,17 @@ async def get_sorted_dicom_images(
             auth=auth,
         )
 
+        instance_id = latest_image[1]
+        acq_datetime = latest_image[0]
+        small_id = consistent_hash(instance_id, 10**8)
+        instance_id = (
+            f"{acq_datetime.strftime('%Y-%m-%d-%H:%M:%S')} ID-{small_id}"
+        )
+
         latest_image = (
-            latest_image[0],
+            acq_datetime,
             file_response.content,
-            latest_image[1],
+            instance_id,
         )
 
         if not final_images_with_dates:
@@ -160,3 +210,23 @@ async def get_sorted_dicom_images(
         (image, instance_id)
         for _, image, instance_id in final_images_with_dates
     ], latest_acq_time
+
+
+class UnauthorizedException(Exception):
+    """Raised when the user is unauthorized and should be redirected."""
+
+    pass
+
+
+def validate_auth_token(auth_token: str):
+    token_data = TOKEN_STORE.get(auth_token)
+    if not token_data or token_data["expires"] < datetime.utcnow():
+        raise UnauthorizedException
+    return token_data["username"]
+
+
+def validate_api_token(api_token: str):
+    token_data = API_TOKEN_STORE.get(api_token)
+    if not token_data or token_data["expires"] < datetime.utcnow():
+        raise HTTPException(status_code=403, detail="Invalid API token.")
+    return token_data["username"]
